+++ educovid-backend/educovid-backend/src/main/java/es/upm/dit/isst/educovid/dao/ResponsableDAOImpl.java
@@ -98,7 +98,7 @@ public class ResponsableDAOImpl implements ResponsableDAO{
 	public ResponsableCOVID readResponsablebyId(String id) {
 		Session session = SessionFactoryService.get().openSession();
 		session.beginTransaction();
-		ResponsableCOVID responsable = session.get(ResponsableCOVID.class, Integer.parseInt(id));
+		ResponsableCOVID responsable = session.get(ResponsableCOVID.class, Long.parseLong(id));
 		session.getTransaction().commit();
 		session.close();
 		return responsable;
